+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ## 1.1.0 Alpha 1 (`1010011`)
 
+- **New:** Seed for media gallery category (#7)
 - **Change:** Set minimum Faker version to 1.9 (#3)
 - **Change:** Require each seed to have its own CLI command to allow better configuration (#5)